+++ CHANGELOG.md
@@ -3,5 +3,10 @@
 This project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.0.1 - 2018-08-13
+## 1.1.0 - 2018-08-15
+### Added
+- Source for new nodes to generate an accurate source map.
+- Warning when no file found for a glob.
+
+## 1.0.0 - 2018-08-13
 - Initial release.